+++ TestPage.cpp
@@ -91,6 +91,16 @@ void TestPage::newTest()
 
 void TestPage::deleteTest()
 {
+  int rc = QMessageBox::warning(this,
+  					    tr("Qtstalker: Warning"),
+					    tr("Are you sure you want to delete this backtest rule?"),
+					    QMessageBox::Yes,
+					    QMessageBox::No,
+					    QMessageBox::NoButton);
+
+  if (rc == QMessageBox::No)
+    return;
+
   config->deleteTest(nav->currentText());
   nav->updateList();
   testNoSelection();
